+++ src/renderer/Renderer.js
@@ -153,7 +153,7 @@ export default class Renderer {
         }
 
         // Assume that all dataframes of a renderLayer share the same metadata
-        const metadata = dataframes.length ? dataframes[0].metadata : {};
+        const metadata = dataframes.length ? dataframes[0].metadata : null;
 
         viewportExpressions.forEach(expr => expr._resetViewportAgg(metadata));