+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add new `clusterCount()` expression
 - Support feature-dependant arrays in combination with ramp
+- Add support to `.blendTo` with String API expressions
 
 ### Fixed
 - Add default export to allow `import carto from '@carto/carto-vl';`